+++ CHANGELOG.md
@@ -23,6 +23,8 @@
   - [#1279](https://github.com/wix-incubator/rich-content/pull/1279) text & highlight color in mobile
 - `image`
   - [#1277](https://github.com/wix-incubator/rich-content/pull/1277) image with link in initial state
+- `viewer`
+  - [#1285](https://github.com/wix-incubator/rich-content/pull/1285) text alignment with punctuation
 </details>
 <hr/>